requirements.txt
asgiref==3.2.10
Django==2.2.8
django-3-jet==1.0.8
django-ckeditor==5.9.0
# django-jet==1.0.8
django-js-asset==1.2.2
djangorestframework==3.11.1
gunicorn==20.0.4
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
